refactor: simplify category feature slice.
