--- lib/features/category/domain/domain.dart
@@ -1,9 +1,0 @@
-// Failures.
-export 'failures/category_add_failed.dart';
-export 'failures/category_edit_failed.dart';
-
-// Services.
-export 'services/add_category.dart';
-export 'services/edit_category.dart';
-export 'services/remove_category.dart';
-export 'services/change_categories_order.dart';
