--- lib/features/category/presentation/presentation.dart
@@ -1,11 +1,0 @@
-// Pages.
-export 'pages/category_page.dart';
-
-// Providers.
-export 'providers/categories_page_provider.dart';
-export 'providers/categories_selection_provider.dart';
-
-// Widgets.
-export 'widgets/add_dialog.dart';
-export 'widgets/category_list.dart';
-export 'widgets/edit_dialog.dart';
