+++ lib/features/category/services/add_category.dart
@@ -3,7 +3,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nacht/common/common.dart';
 import 'package:nacht/core/core.dart';
 import 'package:nacht/database/database.dart';
-import 'package:nacht/features/category/domain/failures/category_add_failed.dart';
+
+import '../failures/category_add_failed.dart';
 
 final addCategoryProvider = Provider<AddCategory>(
   (ref) => AddCategory(
