+++ lib/features/category/services/remove_category.dart
