+++ lib/features/category/views/category_page.dart
@@ -4,7 +4,10 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nacht/core/core.dart';
 import 'package:nacht/widgets/widgets.dart';
 
-import '../presentation.dart';
+import '../providers/categories_page_provider.dart';
+import '../providers/categories_selection_provider.dart';
+import '../widgets/add_dialog.dart';
+import '../widgets/category_list.dart';
 
 class CategoryPage extends ConsumerWidget with LoggerMixin {
   const CategoryPage({Key? key}) : super(key: key);
