+++ lib/features/category/widgets/category_list.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../presentation.dart';
+import '../providers/categories_page_provider.dart';
+import '../providers/categories_selection_provider.dart';
+import 'edit_dialog.dart';
 
 class CategoryList extends HookConsumerWidget {
   const CategoryList({Key? key}) : super(key: key);
