+++ lib/features/category/widgets/edit_dialog.dart
@@ -2,24 +2,36 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:nacht/common/common.dart';
 
-import '../presentation.dart';
+import '../providers/categories_page_provider.dart';
 
-class AddCategoryDialog extends HookConsumerWidget {
-  const AddCategoryDialog({Key? key}) : super(key: key);
+class EditCategoryDialog extends HookConsumerWidget {
+  const EditCategoryDialog({Key? key, required this.categoryData})
+      : super(key: key);
+
+  final CategoryData categoryData;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final notifier = ref.watch(categoriesPageProvider.notifier);
     final controller = useTextEditingController();
 
+    useEffect(() {
+      controller.text = categoryData.name;
+      return null;
+    });
+
     void submit() {
-      notifier.add(controller.value.text);
+      // TODO: Add proper validation.
+      assert(controller.value.text.isNotEmpty);
+
+      notifier.edit(categoryData, controller.value.text);
       context.router.pop();
     }
 
     return AlertDialog(
-      title: const Text('Add category'),
+      title: const Text('Edit category'),
       content: TextField(
         autofocus: true,
         controller: controller,
